Move viewlets to base, use already implemented custom pathbar.
