+++ opengever/base/profiles/default/metadata.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0"?>
 <metadata>
-  <version>4101</version>
+  <version>4200</version>
 </metadata>
