+++ opengever/base/upgrades/profiles/4200/viewlets.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0"?>
+<object>
+    <hidden manager="plone.contentviews" skinname="*">
+        <viewlet name="plone.contentviews" />
+    </hidden>
+    <order manager="plone.contentviews" skinname="*">
+        <viewlet name="opengever.contentviews" insert-before="*" />
+    </order>
+
+    <hidden manager="plone.abovecontent" skinname="*">
+        <viewlet name="plone.path_bar" />
+    </hidden>
+    <order manager="plone.abovecontent" skinname="*">
+        <viewlet name="opengever.path_bar" insert-before="*" />
+    </order>
+</object>
+
