+++ opengever/base/viewlets/configure.zcml
@@ -2,62 +2,71 @@
     xmlns:zope="http://namespaces.zope.org/zope"
     xmlns:browser="http://namespaces.zope.org/browser">
 
-   <browser:viewlet
-       name="plone.belowcontentbody.contenthistory"
-       for="plone.app.versioningbehavior.behaviors.IVersioningSupport"
-       manager="plone.app.layout.viewlets.interfaces.IBelowContentBody"
-       class=".history.DocumentContentHistoryViewlet"
-       template="history.pt"
-       permission="zope2.View"
-       />
+  <browser:viewlet
+     name="plone.belowcontentbody.contenthistory"
+     for="plone.app.versioningbehavior.behaviors.IVersioningSupport"
+     manager="plone.app.layout.viewlets.interfaces.IBelowContentBody"
+     class=".history.DocumentContentHistoryViewlet"
+     template="history.pt"
+     permission="zope2.View"
+     />
 
-    <browser:viewlet
-        name="plone.path_bar"
-        layer="opengever.base.interfaces.IOpengeverBaseLayer"
-        manager="plone.app.layout.viewlets.interfaces.IAboveContent"
-        class=".pathbar.PathBar"
-        permission="zope2.View"
-        />
+  <browser:viewlet
+      name="opengever.path_bar"
+      layer="opengever.base.interfaces.IOpengeverBaseLayer"
+      manager="plone.app.layout.viewlets.interfaces.IAboveContent"
+      class=".pathbar.PathBar"
+      permission="zope2.View"
+      />
 
-   <browser:viewlet
-       name="plone.belowcontenttitle.documentbyline"
-       for="plone.dexterity.interfaces.IDexterityContainer"
-       manager="plone.app.layout.viewlets.interfaces.IBelowContentTitle"
-       class="opengever.base.viewlets.byline.BylineBase"
-       permission="zope2.View"
-       />
+  <browser:viewlet
+     name="plone.belowcontenttitle.documentbyline"
+     for="plone.dexterity.interfaces.IDexterityContainer"
+     manager="plone.app.layout.viewlets.interfaces.IBelowContentTitle"
+     class="opengever.base.viewlets.byline.BylineBase"
+     permission="zope2.View"
+     />
 
-   <browser:viewlet
-       name="plone.belowcontenttitle.documentbyline"
-       for="plone.dexterity.interfaces.IDexterityItem"
-       manager="plone.app.layout.viewlets.interfaces.IBelowContentTitle"
-       class="opengever.base.viewlets.byline.BylineBase"
-       permission="zope2.View"
-       />
+  <browser:viewlet
+     name="plone.belowcontenttitle.documentbyline"
+     for="plone.dexterity.interfaces.IDexterityItem"
+     manager="plone.app.layout.viewlets.interfaces.IBelowContentTitle"
+     class="opengever.base.viewlets.byline.BylineBase"
+     permission="zope2.View"
+     />
 
-   <browser:viewlet
-       name="plone.belowcontenttitle.documentbyline"
-       for="Products.CMFPlone.interfaces.siteroot.IPloneSiteRoot"
-       manager="plone.app.layout.viewlets.interfaces.IBelowContentTitle"
-       class="opengever.base.viewlets.byline.PloneSiteByline"
-       permission="zope2.View"
-       />
+  <browser:viewlet
+     name="plone.belowcontenttitle.documentbyline"
+     for="Products.CMFPlone.interfaces.siteroot.IPloneSiteRoot"
+     manager="plone.app.layout.viewlets.interfaces.IBelowContentTitle"
+     class="opengever.base.viewlets.byline.PloneSiteByline"
+     permission="zope2.View"
+     />
 
-       <browser:viewlet
-           name="gever-colorization"
-           manager="plone.app.layout.viewlets.interfaces.IPortalTop"
-           template="colorization.pt"
-           class=".colorization.ColorizationViewlet"
-           permission="zope2.View"
-           />
+  <browser:viewlet
+     name="gever-colorization"
+     manager="plone.app.layout.viewlets.interfaces.IPortalTop"
+     template="colorization.pt"
+     class=".colorization.ColorizationViewlet"
+     permission="zope2.View"
+     />
 
-   <!-- searchbox viewlet -->
-   <browser:viewlet
-       name="plone.searchbox"
-       manager="plone.app.layout.viewlets.interfaces.IPortalHeader"
-       class=".searchbox.SearchBoxViewlet"
-       permission="zope2.View"
-       layer="opengever.base.interfaces.IOpengeverBaseLayer"
-       />
+  <!-- searchbox viewlet -->
+  <browser:viewlet
+     name="plone.searchbox"
+     manager="plone.app.layout.viewlets.interfaces.IPortalHeader"
+     class=".searchbox.SearchBoxViewlet"
+     permission="zope2.View"
+     layer="opengever.base.interfaces.IOpengeverBaseLayer"
+     />
+
+  <!-- Custom content views -->
+  <browser:viewlet
+      name="opengever.contentviews"
+      manager="plone.app.layout.viewlets.interfaces.IContentViews"
+      permission="zope2.View"
+      layer="opengever.base.interfaces.IOpengeverBaseLayer"
+      class=".contentviews.ModelContentViewsViewlet"
+      />
 
 </configure>
