+++ opengever/meeting/configure.zcml
@@ -1,6 +1,5 @@
 <configure
     xmlns="http://namespaces.zope.org/zope"
-    xmlns:browser="http://namespaces.zope.org/browser"
     xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
     xmlns:grok="http://namespaces.zope.org/grok"
     xmlns:i18n="http://namespaces.zope.org/i18n"
@@ -34,23 +33,5 @@
           id="opengever.meeting.AddCommittee"
           title="opengever.meeting: Add Committee"
           />
-
-    <!-- Custom content views -->
-    <browser:viewlet
-        name="opengever.contentviews"
-        manager="plone.app.layout.viewlets.interfaces.IContentViews"
-        permission="zope2.View"
-        layer="opengever.base.interfaces.IOpengeverBaseLayer"
-        class=".viewlets.ModelContentViewsViewlet"
-        />
-
-    <!-- Custom breadcrumbs -->
-    <browser:viewlet
-        name="opengever.path_bar"
-        manager="plone.app.layout.viewlets.interfaces.IAboveContent"
-        permission="zope2.View"
-        layer="opengever.base.interfaces.IOpengeverBaseLayer"
-        class=".viewlets.ModelPathbarViewlet"
-        />
 
 </configure>
